clean logs, added dotenv to handle env variables
