+++ playwright/playwright.config.ts
@@ -4,6 +4,7 @@ import { defineConfig, devices } from "@playwright/test";
  * See https://playwright.dev/docs/test-configuration.
  */
 export default defineConfig({
+  globalSetup: "./env-setup.ts",
   testDir: "./tests",
   /* Run tests in files in parallel */
   fullyParallel: true,
